[feat] LoggerUtils | Only Show Debug
