+++ Nabi/presentation/src/main/java/com/nabi/nabi/base/NabiApplication.kt
@@ -2,6 +2,8 @@ package com.nabi.nabi.base
 
 import android.app.Application
 import com.kakao.sdk.common.KakaoSdk
+import com.kakao.sdk.common.util.Utility
+import com.nabi.nabi.utils.LoggerUtils
 import com.nabi.nabi.BuildConfig
 import dagger.hilt.android.HiltAndroidApp
 
@@ -16,5 +18,12 @@ class NabiApplication : Application() {
 
         application = this
         KakaoSdk.init(this, BuildConfig.KAKAO_NATIVE_KEY)
+
+        getHashKey()
+    }
+
+    private fun getHashKey(){
+        val keyHash = Utility.getKeyHash(this)
+        LoggerUtils.i("KeyHash: $keyHash")
     }
 }
