+++ Nabi/presentation/src/main/java/com/nabi/nabi/views/diary/add/AddDiarySelectDateFragment.kt
@@ -3,7 +3,7 @@ package com.nabi.nabi.views.diary.add
 import android.annotation.SuppressLint
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.nabi.data.utils.LoggerUtils
+import com.nabi.nabi.utils.LoggerUtils
 import com.nabi.nabi.R
 import com.nabi.nabi.base.BaseFragment
 import com.nabi.nabi.databinding.FragmentSelectDateBinding
