+++ Nabi/presentation/src/main/java/com/nabi/nabi/views/diary/view/SelectDiaryMonthFragment.kt
@@ -3,7 +3,7 @@ package com.nabi.nabi.views.diary.view
 import android.os.Bundle
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.GridLayoutManager
-import com.nabi.data.utils.LoggerUtils
+import com.nabi.nabi.utils.LoggerUtils
 import com.nabi.domain.model.diary.DiaryInfo
 import com.nabi.nabi.R
 import com.nabi.nabi.base.BaseFragment
