+++ Nabi/presentation/src/main/java/com/nabi/nabi/views/home/HomeFragment.kt
@@ -2,7 +2,7 @@ package com.nabi.nabi.views.home
 
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.nabi.data.utils.LoggerUtils
+import com.nabi.nabi.utils.LoggerUtils
 import com.nabi.nabi.R
 import com.nabi.nabi.base.BaseFragment
 import com.nabi.nabi.databinding.FragmentHomeBinding
